LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-align-baseline-001-expected.txt

line1
line2
line1
line2
line1
line2
line1
line2
line1
line2
line1
line2

PASS #target > div 1
PASS #target > div 2
PASS #target > div 3
PASS #target > div 4
FAIL #target > div 5 assert_equals:
<div data-offset-x="35">line1<br>line2</div>
offsetLeft expected 35 but got 15
FAIL #target > div 6 assert_equals:
<div data-offset-x="42">line1<br>line2</div>
offsetLeft expected 42 but got 22
